+++ changelog.md
@@ -1,9 +1,10 @@
 # Changelog
 
 version 5.6.0 -- upcoming (nightly)
-  * custom export templates (see [Features](/docs/features.md))
+  * code templates (see [Features](/docs/features.md))
   * some predefined exporters for
     * Spring RestTemplate
+    * Spring Webclient
 
 version 5.5.0